requirements.txt
django 
djangorestframework
pillow
drf_yasg